some refactoring and adding doc strings
